refactor: split drawdown period series
