+++ scripts/generate_quality_baseline.py
@@ -325,7 +325,7 @@ evidence for PR readiness, not a completion claim.
 | Dimension | Baseline evidence | Current evidence | Improvement shown | Remaining target |
 | --- | --- | --- | --- | --- |
 | API modularity | `src/app/main.py` had 22 route/middleware/handler decorators and 980 lines | `src/app/main.py` has 0 route/middleware/handler decorators and {main_lines} lines | App construction, routers, middleware, errors, and downstream dependency resolution are split into modules | Keep router boundaries green and prevent app-entry-point regression |
-| Code size | Largest files included `src/app/services/concentration_engine.py` at 981 lines and `src/app/main.py` at 980 lines | Largest source files are contract/service modules; no source file over 800 lines after the latest baseline | Monolithic API, concentration service, concentration input/output contracts, concentration metric/response output contracts, concentration issuer mapping, rolling input/output contracts, rolling metric/response output contracts, risk input/output contracts, risk benchmark period metrics, risk period results, risk benchmark metrics, drawdown input/output contracts, drawdown metric/response output contracts, drawdown episode construction, attribution input/output contracts, attribution exposure points, attribution set results, attribution stateful returns, scenario input/output contracts, concentration stateless resolution, concentration simulation resolution, concentration simulation snapshot state, lotus-performance async returns payload handling, attribution decomposition, attribution stateful exposure sourcing, drawdown series math, rolling benchmark metric-series, rolling metric-series, rolling stateful input resolution, rolling stateful source responses, rolling period results, contract example payloads, and OpenAPI request example payloads were split | Continue reducing service and contract hotspots over 600 lines |
+| Code size | Largest files included `src/app/services/concentration_engine.py` at 981 lines and `src/app/main.py` at 980 lines | Largest source files are contract/service modules; no source file over 800 lines after the latest baseline | Monolithic API, concentration service, concentration input/output contracts, concentration metric/response output contracts, concentration issuer mapping, rolling input/output contracts, rolling metric/response output contracts, risk input/output contracts, risk benchmark period metrics, risk period results, risk benchmark metrics, drawdown input/output contracts, drawdown metric/response output contracts, drawdown episode construction, drawdown period-series preparation, attribution input/output contracts, attribution exposure points, attribution set results, attribution stateful returns, scenario input/output contracts, concentration stateless resolution, concentration simulation resolution, concentration simulation snapshot state, lotus-performance async returns payload handling, attribution decomposition, attribution stateful exposure sourcing, drawdown series math, rolling benchmark metric-series, rolling metric-series, rolling stateful input resolution, rolling stateful source responses, rolling period results, contract example payloads, and OpenAPI request example payloads were split | Continue reducing service and contract hotspots over 600 lines |
 | Largest behavior units | Largest function/class included `calculate_risk` at 284 lines, `calculate_rolling_metrics` at 230 lines, and `LotusPerformanceClient` at 256 lines | Largest remaining function is `{largest_function.name}` at {largest_function.lines} lines; `LotusPerformanceClient` is {lotus_performance_client.lines} lines | Large engines and clients were decomposed into helpers, services, routers, and polling/parsing functions | Continue reducing engine-level orchestration hotspots |
 | Complexity | Baseline reported C-or-worse candidates across large service, contract, and readiness code | Current baseline reports no C-or-worse candidates in the complexity snapshot | C-level candidates in concentration parsing, risk period resolution, rolling/attribution validation, and enterprise authorization were removed | Keep radon report-only evidence clean while thresholds are tightened |
 | Architecture enforcement | Import-linter, architecture docs, and quality workflow were introduced as report-only baseline | `make architecture-gate` is green locally and in feature-lane CI | Architecture boundary checks are now part of routine slice validation | Extend contracts as service boundaries mature |
