+++ src/app/services/drawdown_periods.py
@@ -1,8 +1,5 @@
 from __future__ import annotations
 
-from dataclasses import dataclass
-from datetime import date
-from collections.abc import Sequence
 from typing import cast
 
 import pandas as pd
@@ -15,7 +12,12 @@ from app.contracts.drawdown import (
     RelativeDrawdownContext,
     DrawdownStatelessInput,
 )
-from app.contracts.risk import ReturnPoint, RiskRequestPeriod
+from app.services.drawdown_period_series import (
+    DrawdownInputFrames,
+    DrawdownPeriodSeries,
+    build_input_frames,
+    period_series,
+)
 from app.services.drawdown_relative_benchmark import (
     RelativeBenchmarkResult,
     RelativeBenchmarkSeries,
@@ -27,77 +29,6 @@ from app.services.drawdown_series import (
     drawdown_summary as _drawdown_summary,
     to_underwater_series as _to_underwater_series,
 )
-from app.services.risk.period_resolution import resolve_period
-
-
-@dataclass(frozen=True)
-class DrawdownInputFrames:
-    portfolio: pd.DataFrame
-    benchmark: pd.DataFrame
-
-
-@dataclass(frozen=True)
-class DrawdownPeriodSeries:
-    name: str
-    start: date
-    end: date
-    portfolio_returns: pd.Series
-    benchmark_returns: pd.Series
-    benchmark_available: bool
-
-
-def _build_returns_df(returns: Sequence[ReturnPoint]) -> pd.DataFrame:
-    df = pd.DataFrame([{"date": point.date, "value": point.value} for point in returns])
-    if df.empty:
-        return df
-    df["date"] = pd.to_datetime(df["date"])
-    return df.sort_values("date").set_index("date")
-
-
-def _filter_period(df: pd.DataFrame, *, start: date, end: date) -> pd.Series:
-    filtered = df.loc[(df.index >= pd.Timestamp(start)) & (df.index <= pd.Timestamp(end)), "value"]
-    return filtered
-
-
-def _period_name(period: RiskRequestPeriod) -> str:
-    return period.name or period.type
-
-
-def build_input_frames(request: DrawdownStatelessInput) -> DrawdownInputFrames:
-    return DrawdownInputFrames(
-        portfolio=_build_returns_df(request.returns),
-        benchmark=_build_returns_df(request.benchmark_returns),
-    )
-
-
-def _period_series(
-    *,
-    frames: DrawdownInputFrames,
-    request: DrawdownStatelessInput,
-    period: RiskRequestPeriod,
-    open_date: date,
-) -> DrawdownPeriodSeries:
-    start, end = resolve_period(
-        period.type,
-        request.scope.as_of_date,
-        open_date,
-        year=period.year,
-        from_date=period.from_date,
-        to_date=period.to_date,
-    )
-    benchmark_returns = (
-        _filter_period(frames.benchmark, start=start, end=end)
-        if not frames.benchmark.empty
-        else pd.Series(dtype="float64")
-    )
-    return DrawdownPeriodSeries(
-        name=_period_name(period),
-        start=start,
-        end=end,
-        portfolio_returns=_filter_period(frames.portfolio, start=start, end=end),
-        benchmark_returns=benchmark_returns,
-        benchmark_available=not frames.benchmark.empty,
-    )
 
 
 def _insufficient_period_result(
@@ -237,14 +168,14 @@ def drawdown_period_results(
     open_date = cast(pd.Timestamp, frames.portfolio.index.min()).date()
     results: dict[str, DrawdownPeriodResult] = {}
     for period in request.periods:
-        period_series = _period_series(
+        series = period_series(
             frames=frames,
             request=request,
             period=period,
             open_date=open_date,
         )
-        results[period_series.name] = _calculate_period_result(
-            period_series,
+        results[series.name] = _calculate_period_result(
+            series,
             analysis_options=analysis_options,
             include_benchmark=include_benchmark,
         )
